(18) Cambiar entero por float - (19) calculo de hipotenusa - (20) sin terminar

+++ Practica_Ejercicios/18-Main.c
@@ -2,9 +2,9 @@
 #include <math.h>
 
 int main(void){
-    int radioc;
+    float radioc;
     printf("Ingrese el radio del circulo: ");
-    scanf("%d", &radioc);
-    printf("El valor de area a partir del radio es: %f", M_PI*(pow(radioc,2)));
+    scanf("%f", &radioc);
+    printf("El valor de area a partir del radio es: %.2f", M_PI*(pow(radioc,2)));
     return 0;
 }
